apps/localhost_all/requirements.txt
# FastAPI stack
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Config
python-dotenv>=1.0.0

# Logging
colorlog>=6.8.0

# HTTP client
httpx>=0.27.0

# Data validation
pydantic>=2.9.0

# Multipart / upload support
python-multipart>=0.0.9

# Image processing (from draw/)
Pillow>=10.0.0
imageio==2.37.2
imageio-ffmpeg>=0.5.1
numpy>=1.24.0

# Go game processing (from katago/)
sgfmill
chardet

# LINE Bot SDK
line-bot-sdk>=3.5.0

# Google Cloud SDK
google-cloud-storage>=2.18.0
google-cloud-compute>=1.15.0

# OpenAI SDK
openai>=1.0.0

# UUID
uuid
